Added dependency check
cairo_version is the simplest function I could find to import to reduce it's load on the script. There's a way to test if a module exists to import without actually import it but it has some nasty stdout so you then have to import modules to deal with that and it's just not worth it.