Fix crash when calculating Asar file content offsets
Some Electron apps, such as Mattermost Desktop 5.0.0 (#704), contain a
directory named "offset" in their app.asar files, which is treated
in code as a dict of dicts representing files and subdirectories.
Hardcode Tray expects a string for actual offset values before
recalculating them. This fix adds a check to make sure that it only
performs the said action on offset numbers.